# Remove pendingTargets map from graph