@Tatskaari 2021-07-26
Nice! I think working through some of the excess garbage we generate will help with memory fragmentation if we get enough of them in.